Feature: Adding seperate views depending on transaction type 1. Added view for QR scan 2. Separated view for manual entry
